% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamsession_sessions.R
\name{clean_jamsession_name}
\alias{clean_jamsession_name}
\title{Clean jamsession name prior to saving an output file}
\usage{
clean_jamsession_name(
  x,
  remove_pattern = "[-._'\\" /\\\\:]+",
  replacement = "_",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{character} string representing a jam session name.}

\item{remove_pattern}{\code{character} string regular expression pattern
passed to \code{gsub()} to define portions of \code{x} to be replaced.}

\item{replacement}{\code{character} string replacement, recommended to be
a single character.}

\item{...}{additional arguments are passed to \code{gsub()}.}
}
\value{
\code{character} vector after adjusting input 'x'.
}
\description{
Clean jamsession name prior to saving an output file
}
\details{
This function is intended to take an R session name as a character
string, and will replace unwanted characters with a fixed string,
by default underscore \code{"_"}. The \code{remove_pattern} and the
\code{replacement} are configurable if needed.
}
\seealso{
Other jamsession internals: 
\code{\link{fileInfo}()},
\code{\link{list_functions}()},
\code{\link{list_jamsessions}()},
\code{\link{list_objects}()},
\code{\link{print_pkg_description}()}
}
\concept{jamsession internals}
