% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamsession-list-functions.R
\name{grep_functions}
\alias{grep_functions}
\title{search for saved R functions}
\usage{
grep_functions(
  grep_string = ".",
  save_date = ".",
  functions_path = jamsession_paths()$functions,
  ignore.case = TRUE,
  return_df = TRUE,
  most_recent = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{grep_string}{\code{character} string with regular expression pattern,
although usually any text substring will work.
\itemize{
\item Note that input may contain the date as seen in \code{rownames()}
of \code{list_functions()} output.
When \code{grep_string} contains multiple values, they are used
in order to find matching object names, using \code{jamba::provigrep()}.
The order typically has no effect, but can be used to retain the
order of objects returned even when the objects have the same
\code{save_date}.
}}

\item{save_date}{\code{character} string or vector, used as a regular
expression search for matching values in the \code{save_date}
column output from \code{list_objects()}. It can be used to search
for objects with a specific \code{save_date}. Remember to use
\code{most_recent=FALSE} when searching for past versions of
stored R objects.}

\item{ignore.case}{\code{logical} passed to \code{grep()} indicating whether
to ignore upper and lower case, by default \code{ignore.case=TRUE}.}

\item{return_df}{\code{logical} indicating whether to return a \code{data.frame}.}

\item{most_recent}{logical whether to return only the most recent
saved version of each matching object, by default \code{most_recent=TRUE}.
Note that it returns the most recent of \emph{each} matching object.}

\item{verbose}{logical whether to print verbose output}

\item{...}{additional arguments are passed to \code{grepl()}.}

\item{objects_path}{character vector of one or more file paths to
search for saved R objects. When \code{objects_path=NULL}, it uses
the output from \code{jamsession_paths()$objects}.}
}
\value{
one of:
\itemize{
\item When \code{return_df=TRUE} it returns \code{data.frame}
\item Otherwise \code{character} vector.
}
}
\description{
search for saved R functions
}
\details{
This function searches for saved R functions in the appropriate
file paths as defined in \code{jamsession_paths()}, using
\code{jamsession_paths()$functions}, or customized with
argument \code{functions_path}.

It is possible to supply multiple file paths using the
argument \code{functions_path}. All functions are returned,
sorted by \code{save_date}, then by the order of \code{functions_path}.
So a file saved on the same date will be returned by first
entry in \code{functions_path}.

When using \code{most_recent=TRUE}, only the first entry is returned,
using the most recent entry ordered by \code{functions_path}.

For example, \code{refresh_functions(grep_functions("myfunctions"))}
will load the most recent function by \code{save_date} where any
ties prefer the order in \code{functions_path}. Similarly,
\code{refresh_functions("my_favorite_function")} does the same thing,
without the grep_patten search step.
}
\seealso{
Other jamsession functions: 
\code{\link{list_functions_exports}()},
\code{\link{refresh_functions}()}
}
\concept{jamsession functions}
