% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamsession_sessions.R
\name{grep_jamsessions}
\alias{grep_jamsessions}
\title{Search for saved R sessions}
\usage{
grep_jamsessions(
  pattern,
  sessions_path = jamsession_paths()$sessions,
  return_df = TRUE,
  sort_by_date = TRUE,
  ignore.case = TRUE,
  most_recent = TRUE,
  include_attrs = FALSE,
  add_stats = FALSE,
  remove_pattern = NULL,
  replacement = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{pattern}{chracter pattern used with \code{grep()} to match session
names.}

\item{sessions_path}{character vector of one or more file paths to search
for saved R sessions. When \code{NULL}, it uses the output from
\code{jamsession_paths()$sessions}.}

\item{return_df}{logical whether to return a vector (FALSE) or
data.frame (TRUE) with detailed information about each session.}

\item{sort_by_date}{logical whether to sort results by date, as opposed
to the order retrieved from \code{\link{list.files}}.}

\item{ignore.case}{logical sent to \code{\link{grep}}, by default TRUE
which searches in case-insensitive mode.}

\item{most_recent}{logical whether to return only the most recent
saved version of each matching session, by default \code{TRUE}.
This argument is passed to \code{list_jamsessions()}.}

\item{include_attrs}{logical indicating whether to include attributes
\code{"file_path"}, \code{"file_size"}, \code{"save_date"}, and \code{"days_old"},
only used when \code{return_df=FALSE}.}

\item{add_stats}{\code{logical} indicating whether to include summary
stats for each session: \code{number_saved} is the number of past versions
of the session; \code{total_size} is the total file size for all saved
versions.}

\item{remove_pattern, replacement}{arguments passed to
\code{clean_jamsession_name()}.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are passed to \code{list_jamessions()}, and
\code{jamba::provigrep()}, as relevent. If \code{session_prefix} or
\code{session_suffix} should be customized, it should be done here.}
}
\description{
Search for saved R sessions
}
\details{
This function searches for saved R sessions using one or more
text or regular expression patterns.

See \code{list_jamsessions()} for more details about stored sessions.

This function can be chained with \code{load_jamsession()}, for
example \code{load_jamsession(grep_jamsessions("my_project"))},
and it will load the most recently saved R session.
}
\examples{
grep_jamsessions(".");

}
\seealso{
Other jamsession sessions: 
\code{\link{load_jamsession}()},
\code{\link{save_jamsession}()},
\code{\link{show_session_versions}()}
}
\concept{jamsession sessions}
