% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamsession.R
\name{grep_objects}
\alias{grep_objects}
\title{search for saved R objects}
\usage{
grep_objects(
  grep_string = ".",
  save_date = ".",
  objects_path = jamsession_paths()$objects,
  ignore.case = TRUE,
  return_df = TRUE,
  most_recent = TRUE,
  add_stats = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{grep_string}{\code{character} string with regular expression pattern,
although usually any text substring will work. When \code{grep_string}
contains multiple values, they are used in order to find matching
object names, using \code{jamba::provigrep()}. The order typically
has no effect, but can be used to retain the order of objects
returned even when the objects have the same \code{save_date}.}

\item{save_date}{\code{character} string or vector, used as a regular
expression search for matching values in the \code{save_date}
column output from \code{list_objects()}. It can be used to search
for objects with a specific \code{save_date}. Remember to use
\code{most_recent=FALSE} when searching for past versions of
stored R objects.}

\item{objects_path}{character vector of one or more file paths to
search for saved R objects. When \code{objects_path=NULL}, it uses
the output from \code{jamsession_paths()$objects}.}

\item{ignore.case}{\code{logical} passed to \code{grep()} indicating whether
to ignore upper and lower case, by default \code{ignore.case=TRUE}.}

\item{return_df}{\code{logical} indicating whether to return a \code{data.frame}.}

\item{most_recent}{logical whether to return only the most recent
saved version of each matching object, by default \code{most_recent=TRUE}.
Note that it returns the most recent of \emph{each} matching object.}

\item{add_stats}{\code{logical} indicating whether to include summary
stats for each object: \code{number_saved} is the number of past versions
of the object; \code{total_size} is the total file size for all saved
versions.}

\item{verbose}{logical whether to print verbose output}

\item{...}{additional arguments are passed to \code{grepl()}.}
}
\value{
\code{data.frame} when \code{return_df=TRUE}; \verb{character vector}
when \code{return_df=FALSE}. The returned data will contain \code{object}
names that match the \code{grep_string} input.
}
\description{
search for saved R objects
}
\details{
This function searches for saved R objects using one or more
text or regular expression patterns.

It is possible to supply multiple file paths using the
argument \code{objects_path}, in which case all objects are returned
for all directories, then they are sorted by \code{save_date}.
The input \code{objects_path} is stored as an ordered factor, such
that two files with the same \code{save_date} will still be ordered
consistent with the input directories in \code{objects_path}.
In this case, when \code{most_recent=TRUE} there will be one entry
returned per \code{object} name, and the first entry will be the
most recent entry ordered by \code{objects_path}.
In practice, it is recommended to use one \code{objects_path}
location, but this option exists.

For example, \code{load_object(grep_objects("my_favorite_object"))}
will load the most recent object by \code{save_date} where any
ties prefer the order in \code{objects_path}. Similarly,
\code{load_object("my_favorite_object")} does the same thing,
without the grep_patten search step.

See \code{save_object()} for examples.
}
\seealso{
Other jamsession objects: 
\code{\link{list_objects}()},
\code{\link{load_object}()},
\code{\link{save_object}()}
}
\concept{jamsession objects}
