% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamsession-list-functions.R
\name{list_functions}
\alias{list_functions}
\title{list available R functions for jamsession}
\usage{
list_functions(
  functions_path = jamsession_paths()$functions,
  most_recent = FALSE,
  function_prefix = "",
  function_suffix = "_functions[.]R$",
  return_type = c("df", "functions", "list"),
  ...
)
}
\arguments{
\item{functions_path}{\code{character} with default using
\code{jamsession_paths()$functions}, with file paths to search
for matching files with suffix \code{'_functions.R'}.}

\item{most_recent}{\code{logical} default FALSE, whether to return
only the most recent file per unique functions name.}

\item{function_prefix, function_suffix}{\code{character} string used as
regular expression patterns searching files in \code{functions_path}.
Default uses no prefix, and suffix \code{'_functions.R'}.}

\item{return_type}{\code{character} default 'df' with type of object to return.
\itemize{
\item 'df': returns \code{data.frame}
\item 'functions': returns \code{character} vector of functions names.
\item 'list': returns \code{list} with both the above.
}}

\item{...}{additional arguments are ignored.}
}
\value{
\code{data.frame} by default, each row is a matching functions file,
with rownames defined using 'functions' and 'save_date'.
The rows are sorted by 'days_old', 'functions', then 'functions_path'
by default.
\itemize{
\item Output with \code{return_type='functions'} is a \code{character} vector of
functions file names.
\item Output when \code{return_type='list'} is a \code{list} with the two
components above, named 'functions_df' and 'functions', respectively.
}
}
\description{
list available R functions for jamsession
}
\details{
This function lists available R functions for jamsession,
in the given \code{functions_path}.
Available functions are defined by files with suffix
\code{'_functions.R'}.

It is possible to supply multiple file paths using the
argument \code{functions_path}. All functions are returned,
sorted by \code{save_date}, then by the order of \code{functions_path}.
So a file saved on the same date will be returned by first
entry in \code{functions_path}.

When using \code{most_recent=TRUE}, only the first entry is returned
for each unique value in the 'functions' column,
ordered by 'save_date' and 'functions_path'.
}
\examples{
# by default the full list includes all past versions
head(list_functions());

# you can return only the most recent for each function
list_functions(most_recent=TRUE);

}
\seealso{
Other jamsession internals: 
\code{\link{clean_jamsession_name}()},
\code{\link{fileInfo}()},
\code{\link{list_jamsessions}()},
\code{\link{list_objects}()},
\code{\link{print_pkg_description}()}
}
\concept{jamsession internals}
