% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamsession-list-functions.R
\name{list_functions_exports}
\alias{list_functions_exports}
\title{List functions file exports}
\usage{
list_functions_exports(grep_string, debug = FALSE, ...)
}
\arguments{
\item{grep_string}{\code{character} passed to \code{grep_functions()} with
default values except where provided via '...' arguments.}

\item{debug}{\code{logical} default FALSE, whether to print debug information
to illustrate the processing.}

\item{...}{all arguments are passed to \code{grep_functions()}}
}
\value{
\code{list} named by matching \code{functions} files, each
with a \code{character} vector of the matching function or exported
object in the .R file.
}
\description{
List functions file exports, experimental, and dependent upon the
R file using roxygen2 style '@exports'. Any .R files that do not
use roxygen2 comments will return no information.
}
\details{
The process looks for any break in roxygen2 comments in an R file.
Specific steps are described:
\itemize{
\item It splits the lines of each file at the point where a non-roxygen2
line is immediately followed by a roxygen2 line.
\item Within each block, it confirms that there is line that
begins with: \verb{#' @export}. If there is no such line, the block
is ignored.
\item Within remaining blocks, it looks for the first non-roxygen2
line that contains an alphabetic character, which also does not
have a '#' occurring before the alphabetic character.
It returns that one line as the "exported object" for that block.
}

An example of a simple roxygen2 style R function is shown below:

\if{html}{\out{<div class="sourceCode">}}\preformatted{#' Function Name
#'
#' @export
function_name <- function
(argument1, argument2, ...)
\{
   # function code
\}

#' Another Function Name
#'
#' @export
another_function <- function
(argument1, argument2, ...)
\{
   # function code
\}

#' Data object
#' @docType data
"Object_name"
}\if{html}{\out{</div>}}

In the example above, the results would be:

\if{html}{\out{<div class="sourceCode">}}\preformatted{"function_name <- function"
"another_function <- function"
"Object_name"
}\if{html}{\out{</div>}}
}
\seealso{
Other jamsession functions: 
\code{\link{grep_functions}()},
\code{\link{refresh_functions}()}
}
\concept{jamsession functions}
