% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamsession_sessions.R
\name{list_jamsessions}
\alias{list_jamsessions}
\title{List saved R sessions}
\usage{
list_jamsessions(
  sessions_path = jamsession_paths()$sessions,
  most_recent = FALSE,
  session_prefix = "^inProgress_",
  session_suffix = "[.]RData$",
  add_stats = FALSE,
  return_type = c("df", "sessions", "list"),
  ...
)
}
\arguments{
\item{sessions_path}{\code{character} vector of one or more file paths to search
for saved R sessions. When \code{NULL}, it uses the output from
\code{jamsession_paths()$sessions}.}

\item{most_recent}{\code{logical} default FALSE, whether to return only
the most recent entry for each session name.
When \code{most_recent} is \code{FALSE} all saved versions are returned for
all R session names.}

\item{session_prefix, session_suffix}{\code{character} string used as
prefix or suffix when searching each path in \code{sessions_path}
for matching file names.}

\item{add_stats}{\code{logical} indicating whether to include summary
stats for each session: \code{number_saved} is the number of past versions
of the session; \code{total_size} is the total file size for all saved
versions.}

\item{...}{additional parameters are ignored}
}
\value{
\code{data.frame} by default, with sessions per row.
When \code{return_type='sessions'} it returns the unique session
names. When \code{return_type='list'} it returns a \code{list} with both
elements.
}
\description{
List saved R sessions
}
\details{
This function lists available saved R sessions,
usually called by \code{grep_jamsessions()}.

When \code{most_recent=TRUE} it returns only the most recent session
based upon the \code{save_date}.

It is possible to supply multiple file paths using the
argument \code{sessions_path}, in which case all sessions are returned
for all directories, then they are sorted by \code{save_date}.
The input \code{sessions_path} is stored as an ordered factor, such
that two sessions with the same \code{save_date} will still be ordered
consistent with the input directories in \code{sessions_path}.
In this case, when \code{most_recent=TRUE} there will be one entry
returned per \code{object} name, and the first entry will be the
most recent entry ordered by \code{save_date} then \code{sessions_path}.
In practice, it is recommended to use one \code{sessions_path}
location, but this option exists.

For example, \code{jamsession(grep_sessions("my_session"))}
will load the most recent session by \code{save_date} where any
ties prefer the order in \code{sessions_path}. Similarly,
\code{load_jamsession("my_session")} does the same thing,
without the grep_patten search step.
}
\examples{
list_jamsessions();

}
\seealso{
Other jamsession internals: 
\code{\link{clean_jamsession_name}()},
\code{\link{fileInfo}()},
\code{\link{list_functions}()},
\code{\link{list_objects}()},
\code{\link{print_pkg_description}()}
}
\concept{jamsession internals}
