% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamsession.R
\name{list_objects}
\alias{list_objects}
\title{list saved R objects}
\usage{
list_objects(
  objects_path = jamsession_paths()$objects,
  most_recent = FALSE,
  object_prefix = "",
  object_suffix = "[.]RData$",
  add_stats = FALSE,
  ...
)
}
\arguments{
\item{objects_path}{character vector of one or more file paths to
search for saved R objects. When \code{objects_path=NULL}, it uses
the output from \code{jamsession_paths()$objects}.}

\item{most_recent}{logical whether to return only the most recent
saved entry for each session name. When \code{most_recent} is \code{FALSE},
this function returns all saved versions for all R session names.}

\item{object_prefix, object_suffix}{\code{character} string used as
prefix or suffix when searching each path in \code{objects_path}
for matching file names. By default, files are expected
to have file extension \code{.RData}.}

\item{add_stats}{\code{logical} indicating whether to include summary
stats for each object.}

\item{...}{additional arguments are ignored.}
}
\value{
list containing
\describe{
\item{object_df}{\code{data.frame} of R objects}
\item{object}{\code{vector} of matching R object names}
}
}
\description{
list saved R objects
}
\details{
This function lists saved R objects for the given \code{objects_path},
usually called by \code{grep_objects()}.

When \code{most_recent=TRUE} it returns only the most recent version
of each file.

It is possible to supply multiple file paths using the
argument \code{objects_path}, in which case all objects are returned
for all directories, then they are sorted by \code{save_date}.
The input \code{objects_path} is stored as an ordered factor, such
that two files with the same \code{save_date} will still be ordered
consistent with the input directories in \code{objects_path}.
In this case, when \code{most_recent=TRUE} there will be one entry
returned per \code{object} name, and the first entry will be the
most recent entry ordered by \code{save_date} then \code{objects_path}.
In practice, it is recommended to use one \code{objects_path}
location, but this option exists.

For example, \code{load_object(grep_objects("my_favorite_object"))}
will load the most recent object by \code{save_date} where any
ties prefer the order in \code{objects_path}. Similarly,
\code{load_object("my_favorite_object")} does the same thing,
without the grep_patten search step.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# by default the full list includes all past versions
head(list_objects());

# you can return only the most recent for each object
list_objects(most_recent=TRUE);

# add_stats makes sense when most_recent=TRUE
list_objects(most_recent=TRUE, add_stats=TRUE);
\dontshow{\}) # examplesIf}
}
\seealso{
Other jamsession internals: 
\code{\link{clean_jamsession_name}()},
\code{\link{fileInfo}()},
\code{\link{list_functions}()},
\code{\link{list_jamsessions}()},
\code{\link{print_pkg_description}()}
}
\concept{jamsession internals}
