% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamsession_functions.R
\name{print_pkg_description}
\alias{print_pkg_description}
\title{Print package DESCRIPTION}
\usage{
print_pkg_description(pkg)
}
\arguments{
\item{pkg}{\code{character} string indicating a package name.}
}
\description{
Print package DESCRIPTION
}
\details{
This function is a simple wrapper around functions to
find the \code{DESCRIPTION} file for an installed R package,
and print lines to the console.

When the package is not installed, or a temporary R package
is not in the search path, error messages are returned accordingly.
}
\examples{
print_pkg_description("base")

}
\seealso{
Other jamsession internals: 
\code{\link{clean_jamsession_name}()},
\code{\link{fileInfo}()},
\code{\link{list_functions}()},
\code{\link{list_jamsessions}()},
\code{\link{list_objects}()}
}
\concept{jamsession internals}
