% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamsession_sessions.R
\name{save_jamsession}
\alias{save_jamsession}
\title{save an R session}
\usage{
save_jamsession(
  session = get("session", envir = envir),
  assign_session = TRUE,
  save_date = jamba::getDate(),
  do_timestamp = TRUE,
  sessions_path = jamsession_paths()$sessions,
  do_prompt = TRUE,
  do_window_title = TRUE,
  save_history = TRUE,
  envir = globalenv(),
  session_prefix = "inProgress_",
  session_suffix = ".RData",
  save_objectlist = TRUE,
  objectlist_suffix = ".objectlist.txt",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{session}{\code{character} session name used to name the .RData file}

\item{assign_session}{logical whether to assign \code{session} in the
global environment \code{.GlobalEnv}, to reinforce the active session.
It is typically not advised to update the \code{.GlobalEnv} inside a
function, however for typical use of \code{load_jamsession()} the
intended result is to provide an consistent R session, and
\code{session} is part of that expectation.}

\item{do_timestamp}{logical whether to run \code{utils::timestamp()} so the
current time and date are written into the \code{.Rhistory} file.}

\item{sessions_path}{character vector of one or more file paths to search
for saved R sessions. When \code{NULL}, it uses the output from
\code{jamsession_paths()$sessions}.}

\item{do_prompt}{logical whether to run \code{jamba::setPrompt()}
to set the R prompt to include the \code{session}.}

\item{do_window_title}{logical whether to set the default device
window title to include the \code{session}.}

\item{save_history}{logical whether to save the R command history
into an \code{.Rhistory} file. Note that the full available R
command history is saved, but this command history is subject
to platform-specific limitations, and thus may not contain the
entire R command history.}

\item{envir}{R environment to save, by default the active global
environment \code{.GlobalEnv}.}

\item{verbose}{logical whether to print verbose output.}

\item{...}{additional arguments passed to \code{base::save()}}

\item{date}{the date string to use when naming the .RData file.
By default the current date is used, called by \code{jamba::getDate()}.}
}
\description{
save an R session
}
\details{
This function saves the current R session to a \code{.RData} file,
and the R history to a \code{.Rhistory} file. It saves these files
to the first write-accessible path in \code{sessions_path}.

When \code{envir} is defined as something other than \code{globalenv()},
only the R objects stored in that environment are saved.
However the R objects are saved as R objects, with no information
about the source environment.

By default this function will assign the value of \code{session}
to \code{session} in the global environment, unless \code{assign_session=FALSE},
in order to reinforce the active R session name, and which
is shown by the R prompt by default, via calling \code{jamba::setPrompt()}.
These changes are intended as visual reminders of the active R
session.
}
\seealso{
Other jamsession sessions: 
\code{\link{grep_jamsessions}()},
\code{\link{list_jamsessions}()},
\code{\link{load_jamsession}()},
\code{\link{show_session_versions}()}
}
\concept{jamsession sessions}
