% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamsession.R
\name{save_object}
\alias{save_object}
\title{save R object to a central location}
\usage{
save_object(
  object_list,
  save_date = jamba::getDate(),
  objects_path = jamsession_paths()$objects,
  object_notes_list = "",
  do_file_info = FALSE,
  object_suffix = ".RData",
  envir = globalenv(),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object_list}{character vector of object names to save, typically only
one value.}

\item{save_date}{character date string to use when naming the file, typically
the current date is used.}

\item{objects_path}{character vector of one or more file paths to
search for saved R objects. When \code{objects_path=NULL}, it uses
the output from \code{jamsession_paths()$objects}.}

\item{object_notes_list}{character vector of optional notes associated with
the R object. Intended as a crude way to allow describing an object.}

\item{do_file_info}{logical whether to print information about the saved
.RData file, for example including the size.}

\item{envir}{the environment from which to obtain the R object.}
}
\value{
invisible file path loaded
}
\description{
save R object to a central location
}
\details{
This function saves a single R object to an .RData file,
versioned by date using \code{jamba::getDate()},
so it can be discovered and re-used by other R sessions.

When multiple objects should be saved together, the recommended method:
\itemize{
\item create a list object that includes these objects
\item save this list object using \code{save_object()}
}

However, when multiple objects are supplied in \code{object_list},
the object names are concatenated with \code{"-"} and this string
is used to define the saved R object file. For example
\code{save_object(c("df1", "df2"))} will save an R object file
using the string \code{"df1-df2"}. When this file is loaded,
two R objects are loaded into the environment: \code{"df1"} and \code{"df2"}.
While not always ideal, this mechanism may be more convenient,
and it is left for the user to decide which is best.

Note that \code{objects_path} may contain multiple directories, and
when this occurs, the directories are attempted in order. The
R object is saved to the first directory that allows the file
to be saved successfully.
}
\seealso{
Other jamsession objects: 
\code{\link{grep_objects}()},
\code{\link{list_objects}()},
\code{\link{load_object}()}
}
\concept{jamsession objects}
