% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamsession_sessions.R
\name{show_session_versions}
\alias{show_session_versions}
\title{show versions saved for a session}
\usage{
show_session_versions(
  x = get("session", envir = .GlobalEnv),
  sessions_path = jamsession_paths()$sessions,
  most_recent = FALSE,
  return_df = TRUE,
  ...
)
}
\arguments{
\item{most_recent}{\code{logical} whether to return only the most recently
saved entry for each matching session name. If FALSE, it returns all
saved versions of all R sessions.}

\item{...}{additional arguments are passed to \code{grep_jamsessions()}.}

\item{session}{\code{character} indicating the session name.}

\item{session_path}{\code{character} vector of one or more file paths to search
for saved R sessions.}
}
\description{
show versions saved for a session
}
\seealso{
Other jamsession sessions: 
\code{\link{grep_jamsessions}()},
\code{\link{load_jamsession}()},
\code{\link{save_jamsession}()}
}
\concept{jamsession sessions}
