% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-fileinfo.R
\name{fileInfo}
\alias{fileInfo}
\title{file information in data.frame format}
\usage{
fileInfo(fileList, ...)
}
\arguments{
\item{fileList}{\code{character} vector with one or more file paths.}

\item{...}{Additional arguments are ignored.}
}
\value{
\code{data.frame} with file information, including "filename" and "size"
as additional colnames as compared to \code{file.info()} output.
}
\description{
file information in data.frame format
}
\details{
This function is a minor extension to \code{file.info()} in that it
adds the filename as a proper colname, and "size" which contains a text
file size.
}
\seealso{
Other jamsession internals: 
\code{\link{clean_jamsession_name}()},
\code{\link{list_functions}()},
\code{\link{list_jamsessions}()},
\code{\link{list_objects}()},
\code{\link{print_pkg_description}()}
}
\concept{jamsession internals}
