% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamsession_init.R
\name{jamsession_paths}
\alias{jamsession_paths}
\title{jamsession file paths}
\usage{
jamsession_paths(
  sessions = NULL,
  objects = NULL,
  functions = NULL,
  create = FALSE,
  mode = "0755",
  recursive = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{sessions, objects, functions}{\code{character} vector with one or more
file directory path locations, suitable for use by \code{list.files(path=x)}.
When a value is defined, it is used to override the existing
\code{options()} value, so the updated value will be used for all
related jamsession functions. Note that any value \code{""} is removed,
for example \code{sessions=""} is equivalent to \code{sessions=NULL}.}

\item{create}{\code{logical} default FALSE, whether to create a file path
that does not already exist. Only the first path is created
for each of 'sessions', 'objects', 'functions' when none exist.
The directory is created using \code{dir.create()} using the arguments
\code{mode} and \code{recursive}.}

\item{mode, recursive}{arguments passed to \code{dir.create()} when
necessary, by the argument \code{create}.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are passed to \code{dir.create()}.}
}
\value{
\code{list} with named elements
'sessions', 'objects', and 'functions'.
}
\description{
jamsession file paths
}
\details{
This function retrieves the file path for each jamsession type:
\enumerate{
\item 'sessions': the folder where R sessions are saved,
stored in \code{getOption("jam.sessions_path")},
default \code{"~/Projects/R-sessions"}.
\item 'objects': the folder where R objects are saved,
stored in \code{getOption("jam.objects_path")},
default \code{"~/Projects/R-objects"}.
\item 'functions': the folder where R functions are saved,
stored in \code{getOption("jam.functions_path")},
default \code{"~/Projects/R-scripts"}.
}

Each may have multiple paths, and may contain \code{"."} to include
the current working directory.

To use persistent values for your R environment, define
\code{options()} in the \verb{~/.Rprofile} in your HOME directory,
or in the folder of the active project.

To use custom file paths in each R session, use either \verb{~/.Rprofile}
or an appropriate method to define the relevant options.
\itemize{
\item \code{options("jam.sessions_path"="/some/path/sessions")},
\item \code{options("jam.objects_path"="/some/path/objects")},
\item \code{options("jam.functions_path"="/some/path/functions")}.
}
}
\examples{
## Display the current or default paths
c(jamsession_paths(functions=NULL))

## Update the functions path
jamsession_paths(functions="~/Projects/R-functions")

## confirm the functions path has been updated
c(jamsession_paths())
getOption("jam.functions_path")

## Revert the functions path
c(jamsession_paths(functions="~/Projects/R-scripts"))

}
\concept{jamsession utilities}
