% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamsession_sessions.R
\name{load_jamsession}
\alias{load_jamsession}
\title{load an R session}
\usage{
load_jamsession(
  session,
  sessions_path = jamsession_paths()$sessions,
  save_date = NULL,
  envir = globalenv(),
  assign_session = TRUE,
  load_history = TRUE,
  do_window_title = TRUE,
  do_prompt = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{session}{character string corresponding to the \code{"session"},
which is matched directly with the \code{"session"} column output from
\code{grep_jamsessions()}, or \code{data.frame} output from \code{grep_jamsessions()}.}

\item{sessions_path}{character vector of one or more file paths to search
for saved R sessions. When \code{NULL}, it uses the output from
\code{jamsession_paths()$session}. This value is used only when
\code{session} is a character vector, and is passed to
\code{grep_jamsessions()}.}

\item{save_date}{optional character string with a specific date to
use when loading a session. This string is matched with
the \code{"save_date"} returned by \code{list_jamsessions()}.
When \code{save_date} is \code{NULL}, the most recent
session is used, which is the default behavior.}

\item{envir}{\code{environment}, where \code{"environment" \%in\% class(envir)}.
By default \code{envir} uses \code{globalenv()} which loads the R session
into the workspace. This behavior is intended for new R sessions,
to restore the R session to the state it was in when the R session
was saved. If loading into an active R session, \code{envir} should
be defined as a separate environment, in order to prevent
R objects in the R session file from overwriting R objects
with the same name in the local R workspace. For example
\verb{session_env=new.env();load_jamsession("New", envir=session_env)}.}

\item{assign_session}{logical whether to assign \code{session}
to the \code{parent.env()} which is typically the global environment
\code{.GlobalEnv}.}

\item{load_history}{logical whether to load the .Rhistory file for the
R session, if available.}

\item{do_window_title}{logical whether to update the device options
to define the title based upon the \code{session}. Currently when \code{TRUE},
the options \code{quartz.options()} and \code{X11.options()} are defined,
if possible.}

\item{do_prompt}{logical whether to run
\code{jamba::setPrompt()} to update the R prompt to include the
\code{session}.}

\item{verbose}{logical whether to print verbose output.}

\item{...}{additional arguments are passed to \code{jamba::setPrompt()}.}
}
\value{
invisible list of R object names loaded into the environment.
One could use this vector of R object names to learn which R objects
are stored in an R session, by creating a new environment, loading
an R session into that environment, then inspecting the R object
names.
}
\description{
load an R session
}
\details{
This function loads a previously-saved R session, loading the most
recently saved R session by the \code{session}. It also by default
sets the R prompt using \code{jamba::setPrompt()}, to indicate the
\code{session} and help reinforce which R session is in the active
environment.

The default workflow is to load the R session into the R
workspace, that is \code{envir=globalenv()}, because the target use
for \code{load_jamsession()} is to call this function in a new,
empty R session in order to continue work on the saved
R session. In this case is the \code{.Rhistory} file is also
available and \code{load_history=TRUE}, it is loaded as well. Note
that loading R history will overwrite the previous active
R history.

However, when \code{envir} is defined as something other than \code{globalenv()},
then the prompt is not altered, and the R session is
loaded into the specific environment. In this case the R history
is not loaded, even when \code{load_history=TRUE}.
When \code{assign_session=TRUE}
this environment is assigned to a new object in \code{globalenv()}
so that it can be attached with \code{attach()}.
}
\seealso{
Other jamsession sessions: 
\code{\link{grep_jamsessions}()},
\code{\link{list_jamsessions}()},
\code{\link{save_jamsession}()},
\code{\link{show_session_versions}()}
}
\concept{jamsession sessions}
